Start prototyping the new Sponge class
