+++ sponge/__init__.py
@@ -1,1 +1,1 @@
-from .sponge import Sponge
+from .old_sponge import Sponge
